+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * added protection against overriding api with blobs
 * fixed error when server kit received an element with undefined or null attributes
 * added a catch-all path (**) to the routers
+* made action, middleware and watcher blob handlers accept multiple arguments
 
 ## v2.0.1